changed some (a lot) in the file structure (tad more logical now, maybe)
