+++ src/main/java/com/codecool/eshipdiary/security/ApiAuthenticationFilter.java
@@ -1,4 +1,4 @@
-package com.codecool.eshipdiary.filter;
+package com.codecool.eshipdiary.security;
 
 import com.codecool.eshipdiary.service.ApiAuthenticationService;
 import org.slf4j.Logger;
