+++ src/main/java/com/codecool/eshipdiary/security/SecurityEvaluationContextExtension.java
@@ -1,11 +1,11 @@
-package com.codecool.eshipdiary.config;
+package com.codecool.eshipdiary.security;
 
 import org.springframework.data.repository.query.spi.EvaluationContextExtensionSupport;
 import org.springframework.security.access.expression.SecurityExpressionRoot;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 
-class SecurityEvaluationContextExtension extends EvaluationContextExtensionSupport {
+public class SecurityEvaluationContextExtension extends EvaluationContextExtensionSupport {
 
     @Override
     public String getExtensionId() {
