+++ src/main/java/com/codecool/eshipdiary/service/RentalService.java
@@ -6,12 +6,17 @@ import com.codecool.eshipdiary.model.Ship;
 import com.codecool.eshipdiary.model.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 
 @Service
 public class RentalService {
+    @Autowired
+    RentalLogRepositoryService rentalLogRepositoryService;
+
     @Autowired
     ShipRepositoryService shipRepositoryService;
 
@@ -51,7 +56,32 @@ public class RentalService {
                 newLog.setCox(oldLog.getCox());
             }
         }
+    }
 
+    @Transactional
+    public void finalize(RentalLog original, RentalLog finalDetails) {
+        original.setRentalEnd(new Date());
 
+        if (finalDetails.getComment() != null) {
+            original.setComment(finalDetails.getComment());
+            original.setDistance(finalDetails.getDistance());
+        }
+
+        if (finalDetails.getInjuredShip() != null) {
+            Ship injuredShip = finalDetails.getInjuredShip();
+            injuredShip.setActive(false);
+            shipRepositoryService.save(injuredShip);
+            original.setInjuredShip(injuredShip);
+        }
+
+        if (finalDetails.getInjuredOars() != null) {
+            for (Oar oar : finalDetails.getInjuredOars()) {
+                oar.setActive(false);
+                oarRepositoryService.save(oar);
+            }
+            original.setInjuredOars(finalDetails.getInjuredOars());
+        }
+        rentalLogRepositoryService.save(original);
+        setOnWaterForInvolvedItemsIn(original, false);
     }
 }
